Support data classes via RedactedBase
This is a not-so-clever workaround to support data classes via `RedactedBase` base class, which overrides `toString()` as `final` and points to an open `toRedactedString()` function that the kotlin built-in data class `toString()` generator won't conflict with. It'd be ideal if we could somehow set this superclass automagically during compilation.

```kotlin
data class Person(
    val name: String,
    @Redacted val ssn: String
) : RedactedBase()
```